requirements.txt
langchain
langgraph
langchain-openai
python-dotenv
apify-client
pydantic
streamlit
pandas
pytest
pytest-mock